+++ backend/canisters/community/api/src/queries/selected_channel_initial.rs
@@ -1,6 +1,6 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use types::{AccessRules, ChannelId, EventIndex, GroupMember, MessageIndex, TimestampMillis, UserId};
+use types::{AccessRules, ChannelId, EventIndex, GroupMember, MessageIndex, TimestampMillis, UserId, VersionedRules};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
@@ -23,5 +23,7 @@ pub struct SuccessResult {
     pub blocked_users: Vec<UserId>,
     pub invited_users: Vec<UserId>,
     pub pinned_messages: Vec<MessageIndex>,
+    // TODO: remove this field once the website is using `access_rules` instead
     pub rules: AccessRules,
+    pub access_rules: VersionedRules,
 }