+++ backend/canisters/community/api/src/updates/send_message.rs
@@ -2,7 +2,7 @@ use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use types::{
     ChannelId, EventIndex, GroupReplyContext, InvalidPollReason, MessageContentInitial, MessageId, MessageIndex,
-    TimestampMillis, User,
+    TimestampMillis, User, Version,
 };
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -15,6 +15,8 @@ pub struct Args {
     pub replies_to: Option<GroupReplyContext>,
     pub mentioned: Vec<User>,
     pub forwarding: bool,
+    #[serde(default)]
+    pub rules_accepted: Option<Version>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -31,6 +33,7 @@ pub enum Response {
     UserSuspended,
     InvalidRequest(String),
     CommunityFrozen,
+    RulesNotAccepted,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]