+++ backend/canisters/group/impl/src/updates/claim_prize.rs
@@ -66,20 +66,23 @@ fn prepare(args: &Args, state: &mut RuntimeState) -> Result<PrepareResult, Box<R
         let min_visible_event_index = member.min_visible_event_index();
         let user_id = member.user_id;
 
-        let (token, amount) = match state
-            .data
-            .chat
-            .events
-            .reserve_prize(args.message_id, min_visible_event_index, user_id, now)
-        {
-            ReservePrizeResult::AlreadyClaimed => return Err(Box::new(AlreadyClaimed)),
-            ReservePrizeResult::Success(token, amount) => (token, amount),
-            ReservePrizeResult::MessageNotFound => return Err(Box::new(MessageNotFound)),
-            ReservePrizeResult::PrizeFullyClaimed => return Err(Box::new(PrizeFullyClaimed)),
-            ReservePrizeResult::PrizeEnded => return Err(Box::new(PrizeEnded)),
-        };
+        let (token, ledger, amount) =
+            match state
+                .data
+                .chat
+                .events
+                .reserve_prize(args.message_id, min_visible_event_index, user_id, now)
+            {
+                ReservePrizeResult::AlreadyClaimed => return Err(Box::new(AlreadyClaimed)),
+                ReservePrizeResult::Success(t, l, a) => (t, l, a),
+                ReservePrizeResult::MessageNotFound => return Err(Box::new(MessageNotFound)),
+                ReservePrizeResult::PrizeFullyClaimed => return Err(Box::new(PrizeFullyClaimed)),
+                ReservePrizeResult::PrizeEnded => return Err(Box::new(PrizeEnded)),
+            };
 
-        let transaction = create_pending_transaction(token, amount, user_id, now_nanos);
+        let fee = token.fee().unwrap(); // TODO send up the transaction fee when creating the prize message
+
+        let transaction = create_pending_transaction(token, ledger, amount.e8s() as u128, fee, user_id, now_nanos);
 
         Ok(PrepareResult {
             group: state.env.canister_id(),