+++ backend/integration_tests/src/communities/join_channel_tests.rs
@@ -5,7 +5,7 @@ use crate::{client, CanisterIds, TestEnv, User};
 use candid::Principal;
 use ic_test_state_machine_client::StateMachine;
 use std::ops::Deref;
-use types::{ChannelId, CommunityId, MessageContent};
+use types::{ChannelId, CommunityId, MessageContent, Version};
 
 #[test]
 fn join_public_channel_succeeds() {
@@ -73,6 +73,7 @@ fn join_private_channel_fails() {
             community_id,
             channel_id,
             invite_code: None,
+            rules_accepted: Some(Version::zero()),
         },
     );
 
@@ -148,6 +149,7 @@ fn join_community_and_channel_in_single_call_succeeds() {
             community_id,
             channel_id,
             invite_code: None,
+            rules_accepted: Some(Version::zero()),
         },
     );
 
@@ -199,6 +201,7 @@ fn invite_non_community_member_to_channel_succeeds() {
             community_id,
             channel_id,
             invite_code: None,
+            rules_accepted: Some(Version::zero()),
         },
     );