+++ frontend/app/src/components/home/BalanceWithRefresh.svelte
@@ -9,38 +9,39 @@
     const user = client.user;
     const dispatch = createEventDispatcher();
 
-    export let token: string = "ICP";
+    export let ledger: string;
     export let value: bigint;
     export let label: string | undefined = undefined;
     export let minDecimals = 4;
     export let bold = false;
-    export let disabled = false;
     export let toppingUp = false;
     export let showTopUp = false;
     export let refreshing = false;
 
+    $: cryptoLookup = client.cryptoLookup;
+    $: tokenDetails = $cryptoLookup[ledger];
+    $: symbol = tokenDetails.symbol;
+
     $: {
-        if (token) {
+        if (ledger) {
             refresh();
         }
     }
 
     export function refresh() {
-        if (disabled) return;
-
         dispatch("click");
         refreshing = true;
 
         return client
-            .refreshAccountBalance(token, user.userId)
+            .refreshAccountBalance(ledger, user.userId)
             .then((val) => {
                 dispatch("refreshed", val);
             })
             .catch((err) => {
                 const errorMessage = $_("unableToRefreshAccountBalance", {
-                    values: { token },
+                    values: { token: symbol },
                 });
-                client.logError(`Failed to refresh ${token} account balance`, err);
+                client.logError(`Failed to refresh ${symbol} account balance`, err);
                 dispatch("error", errorMessage);
             })
             .finally(() => (refreshing = false));
@@ -55,9 +56,9 @@
     {#if label !== undefined}
         <div class="label">{label}</div>
     {/if}
-    <div class="amount" class:bold>{client.formatTokens(value, minDecimals)}</div>
-    <div class="refresh" class:refreshing class:disabled on:click={refresh}>
-        <Refresh size={"1em"} color={disabled ? "var(--button-disabled)" : "var(--icon-txt)"} />
+    <div class="amount" class:bold>{client.formatTokens(value, minDecimals, tokenDetails.decimals)}</div>
+    <div class="refresh" class:refreshing on:click={refresh}>
+        <Refresh size={"1em"} color={"var(--icon-txt)"} />
     </div>
     {#if showTopUp}
         <div class="top-up" on:click={topUp} title={$_("cryptoAccount.topUp")}>
@@ -81,9 +82,7 @@
         @include font-size(fs-140);
         height: $sp5;
         width: $sp5;
-        &:not(.disabled) {
-            cursor: pointer;
-        }
+        cursor: pointer;
         @include mobile() {
             height: 21.59px;
             width: 21.59px;