+++ frontend/openchat-shared/src/domain/chat/chat.ts
@@ -59,6 +59,7 @@ export interface PlaceholderContent {
 }
 
 export type CryptocurrencyDeposit = {
+    ledger: string;
     token: string;
     amountE8s: bigint;
     feeE8s: bigint;
@@ -80,7 +81,7 @@ export type PendingCryptocurrencyWithdrawal = {
 
 export type CompletedCryptocurrencyWithdrawal = {
     kind: "completed";
-    token: string;
+    ledger: string;
     to: string;
     amountE8s: bigint;
     feeE8s: bigint;
@@ -91,7 +92,7 @@ export type CompletedCryptocurrencyWithdrawal = {
 
 export type FailedCryptocurrencyWithdrawal = {
     kind: "failed";
-    token: string;
+    ledger: string;
     to: string;
     amountE8s: bigint;
     feeE8s: bigint;
@@ -111,7 +112,7 @@ export type CryptocurrencyWithdrawal =
 
 export type CompletedCryptocurrencyTransfer = {
     kind: "completed";
-    token: string;
+    ledger: string;
     recipient: string;
     sender: string;
     amountE8s: bigint;
@@ -134,7 +135,7 @@ export type PendingCryptocurrencyTransfer = {
 
 export type FailedCryptocurrencyTransfer = {
     kind: "failed";
-    token: string;
+    ledger: string;
     recipient: string;
     amountE8s: bigint;
     feeE8s: bigint;
@@ -146,11 +147,6 @@ export type CryptocurrencyTransfer =
     | CompletedCryptocurrencyTransfer
     | PendingCryptocurrencyTransfer
     | FailedCryptocurrencyTransfer;
-
-export type CryptocurrencyTransaction =
-    | CryptocurrencyTransfer
-    | CryptocurrencyWithdrawal
-    | CryptocurrencyDeposit;
 
 export interface CryptocurrencyContent {
     kind: "crypto_content";